+++ resources/views/userprofile/view.blade.php
@@ -68,6 +68,21 @@
             </div>
 
             <div class="card mb-4">
+                <div class="card-header">@lang('userprofile.tfa_authentication')</div>
+                <div class="card-body">
+                    @empty($user->tfa_secret)
+                        @component('components.alert.info')
+                            @lang('userprofile.tfa_authentication_not_enabled')
+                        @endcomponent
+                        <a href="{{ route('userprofile.view2FA') }}" class="btn btn-primary">@icon(check) @lang('app.enable')</a>
+                    @else
+                        <p>@lang('userprofile.tfa_authentication_enabled')</p>
+                        <a href="{{ route('userprofile.view2FA') }}" class="btn btn-primary">@icon(times) @lang('app.disable')</a>
+                    @endempty
+                </div>
+            </div>
+
+            <div class="card mb-4">
                 <div class="card-header">@lang('userprofile.language')</div>
                 <div class="card-body p-0">
                     <div class="list-group list-group-flush">